feat: Docker Compose 部署 + PostgreSQL + OIDC 重构 + Decimal 精度修复
部署架构迁移（Docker Compose）：
- 新增 backend/Dockerfile（Python 3.13-slim，非 root，HEALTHCHECK）
- 新增 docker-compose.yml（backend + postgres:16-alpine，健康检查依赖）
- CI/CD 重写：CI 构建镜像推送 GHCR，前端 rsync，SSH docker compose up
- deploy.sh 简化为：备份 → pull → up -d → 健康检查
- nginx.conf 增加 4 级速率限制 + 安全响应头
- ecosystem.config.cjs 保留供 PM2→Docker 迁移过渡

PostgreSQL 支持：
- config.py 新增 DB_BACKEND=postgres + PG_* 配置项
- build_db_url() 支持 postgresql+asyncpg:// URL 构建
- requirements.txt 新增 asyncpg
- deploy.sh 备份自动检测 DB 类型（pg_dump / cp）

认证重构（移除 casdoor SDK）：
- 新增 app/core/oidc.py：轻量 OIDC 客户端，从 .well-known 获取 JWKS
- auth.py 完全移除 casdoor SDK 依赖，改用 OIDCClient
- 移除 CASDOOR_CERTIFICATE 配置（JWKS 自动获取公钥）
- 前端 auth.ts callback 从 query params 改为 JSON body
- requirements.txt 移除 casdoor>=1.22.0

Decimal/Float 精度修复：
- schemas: TradeRequest.shares/QuoteRequest.shares/ResolveRequest.payout → Decimal
- lmsr.py: quantize_cost/quantize_price 支持 float | Decimal 双入口
- market.py/user.py: avg_price 计算改用 Decimal 直除，移除无意义 float 往返

+++ backend/app/schemas/market.py
@@ -1,4 +1,5 @@
 from datetime import datetime
+from decimal import Decimal
 from typing import List, Optional
 
 from pydantic import BaseModel, Field
@@ -35,7 +36,7 @@ class MarketListItem(BaseModel):
 
 class TradeRequest(BaseModel):
     outcome_id: int
-    shares: float = Field(..., gt=0)
+    shares: Decimal = Field(..., gt=0)
 
 
 class TradeResponse(BaseModel):
@@ -47,7 +48,7 @@ class TradeResponse(BaseModel):
 
 class ResolveRequest(BaseModel):
     winning_outcome_id: int
-    payout: float = Field(..., ge=0)
+    payout: Decimal = Field(..., ge=0)
 
 
 class SettleResult(BaseModel):
@@ -90,7 +91,7 @@ class MarketDetailRead(BaseModel):
 
 class QuoteRequest(BaseModel):
     outcome_id: int
-    shares: float = Field(..., gt=0)
+    shares: Decimal = Field(..., gt=0)
     side: str = Field(..., pattern="^(buy|sell)$")
